// src/api-client/models/conversation-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ParticipantInfo } from './participant-info';

/**
 * 
 * @export
 * @interface ConversationResponse
 */
export interface ConversationResponse {
    /**
     * 
     * @type {string}
     * @memberof ConversationResponse
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ConversationResponse
     */
    'type'?: string;
    /**
     * 
     * @type {string}
     * @memberof ConversationResponse
     */
    'participantsHash'?: string;
    /**
     * 
     * @type {string}
     * @memberof ConversationResponse
     */
    'conversationAvatar'?: string;
    /**
     * 
     * @type {string}
     * @memberof ConversationResponse
     */
    'conversationName'?: string;
    /**
     * 
     * @type {Array<ParticipantInfo>}
     * @memberof ConversationResponse
     */
    'participants'?: Array<ParticipantInfo>;
    /**
     * 
     * @type {string}
     * @memberof ConversationResponse
     */
    'createdDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof ConversationResponse
     */
    'modifiedDate'?: string;
}